metodo ruim

+++ src/functions/request/request_parser/request_parser.c
@@ -251,7 +251,7 @@ int  CwebHttpRequest_parse_http_request(struct CwebHttpRequest *self){
             return MAX_HEADER_SIZE_CODE;
         }
 
-        ssize_t res = recv(self->socket, &raw_entries[i], 1, MSG_WAITALL);
+        int res = recv(self->socket, &raw_entries[i], 1, MSG_WAITALL);
         //ssize_t res = read(self->socket, &raw_entries[i],1);
         //printf("v:%d|char:%c\n",raw_entries[i],raw_entries[i]);
